pyproject.toml
[project]
name = "grocery-discord-bot"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "discord.py>=2.3.2",
]
